+++ src/webapp/src/components/pipelines/ConnectorSettings.vue
@@ -566,7 +566,7 @@ export default {
                 <template v-if="!getIsOfKindHidden(setting.kind)">
                   <div v-if="getIsProtected(setting)" class="control">
                     <a
-                      href="https://meltano.com/developer-tools/environment-variables.html#connector-settings-configuration"
+                      href="https://meltano.com/docs/environment-variables.html#connector-settings-configuration"
                       target="_blank"
                       class="button is-small"
                     >