+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 
 - [#1987](https://gitlab.com/meltano/meltano/-/issues/1987) Restore GitLab and Zendesk data sources in UI
+- [#2005](https://gitlab.com/meltano/meltano/-/issues/2005) Add "Don't see your data source here?" option in UI
 
 ### Fixes